requirements.txt
python-telegram-bot
yfinance
pandas
ta
requests